IndexedDB/idbfactory_deleteDatabase5.htm

<!DOCTYPE html>
<meta charset="utf-8">
<title>IDBFactory.deleteDatabase() - no argument</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://dvcs.w3.org/hg/IndexedDB/raw-file/tip/Overview.html#widl-IDBFactory-deleteDatabase-IDBOpenDBRequest-DOMString-name">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>
<div id="log"></div>
<script>
    test( function() {
        assert_throws({name: 'TypeError'}, function() {
            indexedDB.deleteDatabase();
        });
    }, "IDBFactory.deleteDatabase() - no argument.");
</script>
